Separate generation of getters and setters
It is now possible to generate getters and setters on their own, as it
is sometimes usefull to just generate one of the two.

This has been solved by introducing a new multiselection menu before
selecting the fields for the getter/setter generation. As the java lsp is only
offering a command for generating accessors and not for separatly
generating setters/getters, this was the easiest way I could think of
how to implement the feature.