Add jdt.ls.bundles option to load bundles
To integrate nvim-jdtls/nvim-dap with coc.nvim, bundles from [1][2] need
to be loaded.

At the moment this is only possible via CoC extension's javaExtensions,
but it's quite an overkill to create new extensions for this since the
rest of the code is written in Lua.

So adding a new option to allow users to do this themselves.

[1] https://github.com/microsoft/java-debug
[2] https://github.com/microsoft/vscode-java-test